+++ packages/components/src/components/radio/radio.ts
@@ -1,4 +1,3 @@
-import '../icon/icon';
 import { css, html } from 'lit';
 import { customElement } from '../../internal/register-custom-element';
 import { property, state } from 'lit/decorators.js';