+++ packages/components/src/components/select/select.ts
@@ -1,3 +1,5 @@
+import '../icon/icon';
+import '../tag/tag';
 import { animateTo, stopAnimations } from '../../internal/animate.js';
 import { css, html } from 'lit';
 import { customElement } from '../../internal/register-custom-element';
@@ -12,9 +14,8 @@ import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import { waitForEvent } from '../../internal/event.js';
 import { watch } from '../../internal/watch.js';
 import cx from 'classix';
-import SdIcon from '../icon/icon';
+// eslint-disable-next-line
 import SdPopup from '../popup/popup';
-import SdTag from '../tag/tag';
 import SolidElement from '../../internal/solid-element';
 import type { SolidFormControl } from '../../internal/solid-element';
 import type { TemplateResult } from 'lit';
@@ -67,12 +68,6 @@ import type SdOption from '../option/option';
  */
 @customElement('sd-select')
 export default class SdSelect extends SolidElement implements SolidFormControl {
-  static dependencies = {
-    'sd-icon': SdIcon,
-    'sd-popup': SdPopup,
-    'sd-tag': SdTag
-  };
-
   private readonly formControlController: FormControlController = new FormControlController(this, {
     assumeInteractionOn: ['sd-blur', 'sd-input']
   });