+++ packages/docs/src/stories/templates/pagination.stories.ts
@@ -100,6 +100,14 @@ export const Number = {
 
           state.current--;
           update();
+
+          /** 
+           * NOTE: Due to a11y purposes, when the start or end is reached,
+           * the focus must switch to the oposite button
+           */
+          if (state.current === 1) {
+            next.focus();
+          }
         }
 
         const handleNext = (event) => {
@@ -109,6 +117,14 @@ export const Number = {
 
           state.current++;
           update();
+
+          /**
+           * NOTE: Due to a11y purposes, when the start or end is reached,
+           * the focus must switch to the oposite button
+           */
+          if (state.current === pages.length) {
+            previous.focus();
+          }
         }
 
         pages.forEach((page) => page.addEventListener('click', handlePageClick));
@@ -173,6 +189,14 @@ export const Simple = {
 
           state.current--;
           update();
+          
+          /**
+           * NOTE: Due to a11y purposes, when the start or end is reached,
+           * the focus must switch to the oposite button
+           */ 
+          if (state.current === 1) {
+            next.focus();
+          }
         }
 
         const handleNext = (event) => {
@@ -182,6 +206,14 @@ export const Simple = {
 
           state.current++;
           update();
+
+          /**
+           * NOTE: Due to a11y purposes, when the start or end is reached,
+           * the focus must switch to the oposite button
+           */
+          if (state.current === state.total) {
+            previous.focus();
+          }
         }
 
         previous.addEventListener('click', handlePrevious);