+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,69 @@
 # @solid-design-system/docs
 
+## 1.14.0
+
+### Minor Changes
+
+- Introduce overview page for components and styles. _[`#2024`](https://github.com/solid-design-system/solid/pull/2024) [`6fa86c8`](https://github.com/solid-design-system/solid/commit/6fa86c819208ecc26d9ab1394af81e3467aab191) [@smfonseca](https://github.com/smfonseca)_
+
+  These pages include information for:
+
+  - Common use cases
+  - Usage guidelines
+  - Accessibility
+
+  Docs pages are now used only to document technical information, all other information has now been moved to the "overview" page.
+
+## 1.13.11
+
+### Patch Changes
+
+- Added `sd-dialog` new known browser issue. _[`#2140`](https://github.com/solid-design-system/solid/pull/2140) [`034046c`](https://github.com/solid-design-system/solid/commit/034046c54cc0455dc159a2c00a683138e34874f7) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+  Implemented new `Prevent Closing` story on `sd-dialog` templates.
+
+## 1.13.10
+
+### Patch Changes
+
+- Improved `sd-drawer` `nav` aria labels on the drawer template. _[`#2143`](https://github.com/solid-design-system/solid/pull/2143) [`79644ba`](https://github.com/solid-design-system/solid/commit/79644baba6ecba34ac91abd088c814990b6c1b05) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+## 1.13.9
+
+### Patch Changes
+
+- Fix `sd-badge` default story. _[`#2144`](https://github.com/solid-design-system/solid/pull/2144) [`ee96458`](https://github.com/solid-design-system/solid/commit/ee96458e7a9a2f6101e5e455d2d37d29b6122135) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+## 1.13.8
+
+### Patch Changes
+
+- Implemented `Visually Disabled` story on `sd-tab.` _[`#2126`](https://github.com/solid-design-system/solid/pull/2126) [`41c40fd`](https://github.com/solid-design-system/solid/commit/41c40fdfb3da8b4dbe28a357affe922221f3733d) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+## 1.13.7
+
+### Patch Changes
+
+- Improve table template a11y. _[`#2137`](https://github.com/solid-design-system/solid/pull/2137) [`3dbb5ca`](https://github.com/solid-design-system/solid/commit/3dbb5cad431425333b484068cfbb12db6969db8b) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+## 1.13.6
+
+### Patch Changes
+
+- Added new `sd-scrollable` screenshot test named "Auto Orientation". _[`#2109`](https://github.com/solid-design-system/solid/pull/2109) [`d2a3096`](https://github.com/solid-design-system/solid/commit/d2a3096319ca9f55428d24d8d333daded7227a4e) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+## 1.13.5
+
+### Patch Changes
+
+- Added missing `sd-input` `aria-label` on forms template. _[`#2119`](https://github.com/solid-design-system/solid/pull/2119) [`59b8c84`](https://github.com/solid-design-system/solid/commit/59b8c840051167c2b55a94d562536934018d9707) [@paulovareiro29](https://github.com/paulovareiro29)_
+
+## 1.13.4
+
+### Patch Changes
+
+- Improve `sd-pagination` a11y by moving focus to the other previous/next button when the first/last page is reached. _[`#2107`](https://github.com/solid-design-system/solid/pull/2107) [`4b225e7`](https://github.com/solid-design-system/solid/commit/4b225e71e3382f9eda7a2663cd679bd930ddee52) [@paulovareiro29](https://github.com/paulovareiro29)_
+
 ## 1.13.3
 
 ### Patch Changes